+++ fnos/__init__.py
@@ -17,7 +17,7 @@ Fnos - A Python client for Fnos WebSocket communication
 """
 
 from .client import FnosClient
-from .exceptions import NotConnectedError
+from .exceptions import HTTPSRequiredError, NotConnectedError
 from .store import Store
 from .resource_monitor import ResourceMonitor
 from .sac import SAC
@@ -44,6 +44,7 @@ __version__ = "0.13.0"
 
 __all__ = [
     "FnosClient",
+    "HTTPSRequiredError",
     "Store",
     "ResourceMonitor",
     "SAC",